base: Extract widget infrastructure to base.js and mounts.js
Move ElementBase, Chart, TipBox/TipMenu/TipItem, smStyleManager,
RotateBinLayout, color helpers, and Clutter monkey-patches from
extension.js into base.js (~775 lines).

Move smMountsMonitor, Graph, Bar, Pie disk visualization into
mounts.js (~340 lines).

build_menu_info goes in base.js (not extension.js) to avoid circular
dependency with Battery, which calls it directly.

No behavior changes — pure code motion.
