+++ system-monitor-next@paradoxxx.zero.gmail.com/extension.js
@@ -22,8 +22,6 @@ import { Extension, gettext as _ } from "resource:///org/gnome/shell/extensions/
 
 import Clutter from "gi://Clutter";
 import GLib from "gi://GLib";
-import GObject from "gi://GObject";
-import Cogl from "gi://Cogl";
 
 import Gio from "gi://Gio";
 import Shell from "gi://Shell";
@@ -32,1096 +30,28 @@ import UPowerGlib from "gi://UPowerGlib";
 import GTop from "gi://GTop";
 import NM from "gi://NM";
 
-import * as ExtensionSystem from "resource:///org/gnome/shell/ui/extensionSystem.js";
-
 import * as Main from "resource:///org/gnome/shell/ui/main.js";
 import * as PanelMenu from "resource:///org/gnome/shell/ui/panelMenu.js";
 import * as PopupMenu from "resource:///org/gnome/shell/ui/popupMenu.js";
 
-import * as Util from "resource:///org/gnome/shell/misc/util.js";
-
 import { sm_log } from './utils.js';
 import { parse_bytearray, check_sensors } from './common.js';
 import { migrateSettings } from './migration.js';
+import {
+    color_from_string, try_read_int_file,
+    smStyleManager, ElementBase, build_menu_info,
+} from './base.js';
+import { smMountsMonitor, Bar, Pie } from './mounts.js';
 
 const NetworkManager = NM;
 const UPower = UPowerGlib;
 // Copied as of https://gitlab.gnome.org/GNOME/gnome-shell/-/blob/5fa08fe53376f5dca755360bd005a4a51ca78917/js/ui/panel.js#L45
 const PANEL_ICON_SIZE = 16;
 
-// stale network shares will cause the shell to freeze, enable this with caution
-const ENABLE_NETWORK_DISK_USAGE = false;
-
-Clutter.Actor.prototype.raise_top = function raise_top() {
-    const parent = this.get_parent();
-    if (!parent) {
-        return;
-    }
-    parent.set_child_above_sibling(this, null);
-}
-Clutter.Actor.prototype.reparent = function reparent(newParent) {
-    const parent = this.get_parent();
-    if (parent) {
-        parent.remove_child(this);
-    }
-    newParent.add_child(this);
-}
-
-function l_limit(t) {
-    return (t > 0) ? t : 1000;
-}
-
-function change_text() {
-    this.label.visible = this.extension._Schema.get_boolean(this.elt + '-show-text');
-}
-
-function change_style() {
-    let style = this.extension._Schema.get_string(this.elt + '-style');
-    this.text_box.visible = style === 'digit' || style === 'both';
-    this.chart.actor.visible = style === 'graph' || style === 'both';
-}
-
-function build_menu_info(extension) {
-    let elts = extension.__sm.elts;
-    let tray_menu = extension.__sm.tray.menu;
-
-    if (tray_menu._getMenuItems().length &&
-        typeof tray_menu._getMenuItems()[0].actor.get_last_child() !== 'undefined') {
-        tray_menu._getMenuItems()[0].actor.get_last_child().destroy_all_children();
-        for (let elt in elts) {
-            elts[elt].menu_items = elts[elt].create_menu_items();
-        }
-    } else {
-        return;
-    }
-
-    let menu_info_box_table = new St.Widget({
-        style: 'padding: 10px 0px 10px 0px; spacing-rows: 10px; spacing-columns: 15px;',
-        layout_manager: new Clutter.GridLayout({orientation: Clutter.Orientation.VERTICAL})
-    });
-    let menu_info_box_table_layout = menu_info_box_table.layout_manager;
-
-    // Populate Table
-    let row_index = 0;
-    for (let elt in elts) {
-        if (!elts[elt].menu_visible) {
-            continue;
-        }
-
-        // Add item name to table
-        menu_info_box_table_layout.attach(
-            new St.Label({
-                text: elts[elt].item_name,
-                style_class: extension._Style.get('sm-title'),
-                x_align: Clutter.ActorAlign.START,
-                y_align: Clutter.ActorAlign.CENTER
-            }), 0, row_index, 1, 1);
-
-        // Add item data to table
-        let col_index = 1;
-        for (let item in elts[elt].menu_items) {
-            menu_info_box_table_layout.attach(
-                elts[elt].menu_items[item], col_index, row_index, 1, 1);
-
-            col_index++;
-        }
-
-        row_index++;
-    }
-    tray_menu._getMenuItems()[0].actor.get_last_child().add_child(menu_info_box_table);
-}
-
-function change_menu() {
-    this.menu_visible = this.extension._Schema.get_boolean(this.elt + '-show-menu');
-    build_menu_info(this.extension);
-}
-
 function change_usage(extension) {
     let usage = extension._Schema.get_string('disk-usage-style');
     extension.__sm.pie.show(usage === 'pie');
     extension.__sm.bar.show(usage === 'bar');
-}
-
-function cogl_color_from_string(colorString) {
-    let [ok, color] = Cogl.Color.from_string(colorString);
-    if (!ok) {
-        sm_log(`Failed to parse color string (${colorString}). Falling back to red.`);
-        color = new Cogl.Color();
-        Cogl.Color.init_from_4ub(color, 255, 0, 0, 255);
-    }
-    return color;
-}
-
-function clutter_color_from_string(colorString) {
-    return Clutter.Color.from_string(colorString)[1];
-}
-
-function color_from_string(colorString) {
-    if (Cogl.Color.from_string)
-        return cogl_color_from_string(colorString);
-    return clutter_color_from_string(colorString);
-}
-
-function interesting_mountpoint(mount) {
-    if (mount.length < 3) {
-        return false;
-    }
-
-    return ((mount[0].indexOf('/dev/') === 0 || mount[2].toLowerCase() === 'nfs') && mount[2].toLowerCase() !== 'udf');
-}
-
-// Gnome 45 Clutter.cairo_set_source_color compatibility
-function sm_cairo_set_source_color(cr, bg_color) {
-    if (Clutter.cairo_set_source_color)
-        Clutter.cairo_set_source_color(cr, bg_color);
-    else
-        cr.setSourceColor(bg_color);
-}
-
-function try_read_int_file(filename, callback) {
-    if (filename && GLib.file_test(filename, GLib.FileTest.EXISTS)) {
-        let file = Gio.file_new_for_path(filename);
-        file.load_contents_async(null, (source, result) => {
-            let as_r = source.load_contents_finish(result);
-            callback(parseInt(parse_bytearray(as_r[1])));
-        });
-        return true;
-    }
-}
-
-// This is the algorithm used by the df utility. Returns an object with
-// used and total fields, computed from a given statfs structure.
-function calc_usage(statfs) {
-    // bfree represents the total amount of disk space remaining for the
-    // superuser and internal FS operations, while bavail represents the space
-    // remaining for an unprivileged user. The difference between bfree and
-    // bavail represents reserved blocks that should not be part of the total
-    // value, since users don't get to use those blocks. That is one way to
-    // explain why total here is blocks - (bfree - bavail). Alternate
-    // explanation: as bavail approaches 0, used and total should converge.
-    const used = statfs.blocks - statfs.bfree;
-    return {used, total: used + statfs.bavail};
-}
-
-const smStyleManager = class SystemMonitor_smStyleManager {
-    constructor(extension) {
-        this.extension = extension;
-        this._suffix = '';
-        this._iconsize = 1;
-        this._diskunits = _('MiB/s');
-        this._netunits_kbytes = _('KiB/s');
-        this._netunits_mbytes = _('MiB/s');
-        this._netunits_gbytes = _('GiB/s');
-        this._netunits_kbits = _('kbit/s');
-        this._netunits_mbits = _('Mbit/s');
-        this._netunits_gbits = _('Gbit/s');
-        this._pie_size = 300;
-        this._pie_fontsize = 14;
-        this._bar_width = 300;
-
-        this._bar_thickness = 15;
-        this._bar_fontsize = 14;
-        this._compact = this.extension._Schema.get_boolean('compact-display');
-
-        if (this._compact) {
-            this._suffix = '-compact';
-            this._iconsize = 3 / 5;
-            this._diskunits = _('MB');
-            this._netunits_kbytes = _('kB');
-            this._netunits_mbytes = _('MB');
-            this._netunits_gbytes = _('GB');
-            this._netunits_kbits = 'kb';
-            this._netunits_mbits = 'Mb';
-            this._netunits_gbits = 'Gb';
-            this._pie_size *= 4 / 5;
-            this._pie_fontsize = 12;
-            this._bar_width *= 3 / 5;
-            this._bar_thickness = 12;
-            this._bar_fontsize = 12;
-        }
-    }
-    get(style) {
-        return style + this._suffix;
-    }
-    iconsize() {
-        return this._iconsize;
-    }
-    diskunits() {
-        return this._diskunits;
-    }
-    netunits_kbytes() {
-        return this._netunits_kbytes;
-    }
-    netunits_mbytes() {
-        return this._netunits_mbytes;
-    }
-    netunits_gbytes() {
-        return this._netunits_gbytes;
-    }
-    netunits_kbits() {
-        return this._netunits_kbits;
-    }
-    netunits_mbits() {
-        return this._netunits_mbits;
-    }
-    netunits_gbits() {
-        return this._netunits_gbits;
-    }
-    pie_size() {
-        return this._pie_size;
-    }
-    pie_fontsize() {
-        return this._pie_fontsize;
-    }
-    bar_width() {
-        return this._bar_width;
-    }
-    bar_thickness() {
-        return this._bar_thickness;
-    }
-    bar_fontsize() {
-        return this._bar_fontsize;
-    }
-}
-
-const Chart = class SystemMonitor_Chart {
-    constructor(extension, width, height, parent) {
-        this.extension = extension;
-        this.actor = new St.DrawingArea({style_class: extension._Style.get('sm-chart'), reactive: false});
-        this.parentC = parent;
-        this.width = width;
-        let themeContext = St.ThemeContext.get_for_stage(global.stage);
-        this.scale_factor = themeContext.scale_factor;
-        this.actor.set_width(this.width * this.scale_factor);
-        this.actor.set_height(height);
-        this.data = [];
-        for (let i = 0; i < this.parentC.colors.length; i++) {
-            this.data[i] = [];
-        }
-        themeContext.connect('notify::scale-factor', this.rescale.bind(this));
-        this.actor.connect('repaint', this._draw.bind(this));
-    }
-    update() {
-        let data_a = this.parentC.vals;
-        if (data_a.length !== this.parentC.colors.length) {
-            return;
-        }
-        let accdata = [];
-        for (let l = 0; l < data_a.length; l++) {
-            accdata[l] = (l === 0) ? data_a[0] : accdata[l - 1] + ((data_a[l] > 0) ? data_a[l] : 0);
-            this.data[l].push(accdata[l]);
-            if (this.data[l].length > this.width) {
-                this.data[l].shift();
-            }
-        }
-        if (!this.actor.visible) {
-            return;
-        }
-        this.actor.queue_repaint();
-    }
-    _draw() {
-        if (!this.actor.visible) {
-            return;
-        }
-        let [width, height] = this.actor.get_surface_size();
-        let cr = this.actor.get_context();
-        let min = 0, max;
-        if (this.parentC.min) {
-            min = this.parentC.min;
-        }
-        if (this.parentC.max) {
-            max = this.parentC.max;
-        } else {
-            max = Math.max.apply(this, this.data[this.data.length - 1]);
-            max = Math.max(1, Math.pow(2, Math.ceil(Math.log(max) / Math.log(2))));
-            if (this.parentC.graph_scale_cooldown_delay_minutes !== 0) {
-                if (max > this.parentC.graph_scale_max_including_cooldown) {
-                    // Restart the cooldown period with this new max.
-                    const oldMax = this.parentC.graph_scale_max_including_cooldown;
-                    this.parentC.restart_cooldown_timer(max);
-                }
-                this.parentC.graph_scale_max_including_cooldown =
-                    Math.max(max, this.parentC.graph_scale_max_including_cooldown);
-                max = this.parentC.graph_scale_max_including_cooldown;
-            }
-        }
-        const range = max - min, top = 1 + min / range;
-        sm_cairo_set_source_color(cr, this.extension._Background);
-        cr.rectangle(0, 0, width, height);
-        cr.fill();
-        for (let i = this.parentC.colors.length - 1; i >= 0; i--) {
-            let samples = this.data[i].length - 1;
-            if (samples > 0) {
-                cr.moveTo(width, height); // bottom right
-                let x = width - 0.25 * this.scale_factor;
-                cr.lineTo(x, (top - this.data[i][samples] / range) * height);
-                x -= 0.5 * this.scale_factor;
-                for (let j = samples; j >= 0; j--) {
-                    let y = (top - this.data[i][j] / range) * height;
-                    cr.lineTo(x, y);
-                    x -= 0.5 * this.scale_factor;
-                    cr.lineTo(x, y);
-                    x -= 0.5 * this.scale_factor;
-                }
-                x += 0.25 * this.scale_factor;
-                cr.lineTo(x, (top - this.data[i][0] / range) * height);
-                cr.lineTo(x, height);
-                cr.closePath();
-                sm_cairo_set_source_color(cr, this.parentC.colors[i]);
-                cr.fill();
-            }
-        }
-        cr.$dispose();
-    }
-    resize(width) {
-        if (this.width === width) {
-            return;
-        }
-        this.width = width;
-        if (this.width < this.data[0].length) {
-            for (let i = 0; i < this.parentC.colors.length; i++) {
-                this.data[i] = this.data[i].slice(-this.width);
-            }
-        }
-        this.actor.set_width(this.width * this.scale_factor); // repaints
-    }
-    rescale(themeContext) {
-        this.scale_factor = themeContext.scale_factor;
-        this.actor.set_width(this.width * this.scale_factor); // repaints
-    }
-}
-
-// Class to deal with volumes insertion / ejection
-const smMountsMonitor = class SystemMonitor_smMountsMonitor {
-    constructor() {
-        this.files = [];
-        this.num_mounts = -1;
-        this.listeners = [];
-        this.connected = false;
-
-        this._volumeMonitor = Gio.VolumeMonitor.get();
-        let sys_mounts = ['/home', '/tmp', '/boot', '/usr', '/usr/local'];
-        this.base_mounts = ['/'];
-        sys_mounts.forEach((sMount) => {
-            if (this.is_sys_mount(sMount + '/')) {
-                this.base_mounts.push(sMount);
-            }
-        });
-        this.connect();
-    }
-    refresh() {
-        // try check that number of volumes has changed
-        // try {
-        //     let num_mounts = this.manager.getMounts().length;
-        //     if (num_mounts == this.num_mounts)
-        //         return;
-        //     this.num_mounts = num_mounts;
-        // } catch (e) {};
-
-        // Can't get mountlist:
-        // GTop.glibtop_get_mountlist
-        // Error: No symbol 'glibtop_get_mountlist' in namespace 'GTop'
-        // Getting it with mtab
-        // let mount_lines = Shell.get_file_contents_utf8_sync('/etc/mtab').split("\n");
-        // this.mounts = [];
-        // for(let mount_line in mount_lines) {
-        //     let mount = mount_lines[mount_line].split(" ");
-        //     if(interesting_mountpoint(mount) && this.mounts.indexOf(mount[1]) < 0) {
-        //         this.mounts.push(mount[1]);
-        //     }
-        // }
-        // log("[System monitor] old mounts: " + this.mounts);
-        this.mounts = [];
-        for (let base in this.base_mounts) {
-            // log("[System monitor] " + this.base_mounts[base]);
-            this.mounts.push(this.base_mounts[base]);
-        }
-        let mount_lines = this._volumeMonitor.get_mounts();
-        mount_lines.forEach((mount) => {
-            if ((!this.is_net_mount(mount) || ENABLE_NETWORK_DISK_USAGE) &&
-                 !this.is_ro_mount(mount)) {
-                let mpath = mount.get_root().get_path() || mount.get_default_location().get_path();
-                if (mpath) {
-                    this.mounts.push(mpath);
-                }
-            }
-        });
-        // log("[System monitor] base: " + this.base_mounts);
-        // log("[System monitor] mounts: " + this.mounts);
-        for (let i in this.listeners) {
-            this.listeners[i](this.mounts);
-        }
-    }
-    add_listener(cb) {
-        this.listeners.push(cb);
-    }
-    remove_listener(cb) {
-        this.listeners.pop(cb);
-    }
-    get_mounts() {
-        return this.mounts;
-    }
-    is_sys_mount(mpath) {
-        let file = Gio.file_new_for_path(mpath);
-        try {
-            let info = file.query_info(Gio.FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT,
-                Gio.FileQueryInfoFlags.NONE, null);
-            return info.get_attribute_boolean(Gio.FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT);
-        } catch (e) {
-            if (e.matches(Gio.IOErrorEnum, Gio.IOErrorEnum.NOT_FOUND)) {
-                return false;
-            }
-            throw e;
-        }
-    }
-    is_ro_mount(mount) {
-        // FIXME: running this function after "login after waking from suspend"
-        // can make login hang. Actual issue seems to occur when a former net
-        // mount got broken (e.g. due to a VPN connection terminated or
-        // otherwise broken connection)
-        try {
-            let file = mount.get_default_location();
-            let info = file.query_filesystem_info(Gio.FILE_ATTRIBUTE_FILESYSTEM_READONLY, null);
-            return info.get_attribute_boolean(Gio.FILE_ATTRIBUTE_FILESYSTEM_READONLY);
-        } catch (e) {
-            return false;
-        }
-    }
-    is_net_mount(mount) {
-        try {
-            let file = mount.get_default_location();
-            let info = file.query_filesystem_info(Gio.FILE_ATTRIBUTE_FILESYSTEM_TYPE, null);
-            let result = info.get_attribute_string(Gio.FILE_ATTRIBUTE_FILESYSTEM_TYPE);
-            let net_fs = ['nfs', 'smbfs', 'cifs', 'ftp', 'sshfs', 'sftp', 'mtp', 'mtpfs'];
-            return !file.is_native() || net_fs.indexOf(result) > -1;
-        } catch (e) {
-            return false;
-        }
-    }
-    connect() {
-        if (this.connected) {
-            return;
-        }
-        try {
-            this.manager = this._volumeMonitor;
-            this.mount_added_id = this.manager.connect('mount-added', this.refresh.bind(this));
-            this.mount_removed_id = this.manager.connect('mount-removed', this.refresh.bind(this));
-            // need to add the other signals here
-            this.connected = true;
-        } catch (e) {
-            sm_log('Failed to register on placesManager notifications', 'error');
-            sm_log('Got exception : ' + e, 'error');
-        }
-        this.refresh();
-    }
-    disconnect() {
-        if (!this.connected) {
-            return;
-        }
-        this.manager.disconnect(this.mount_added_id);
-        this.manager.disconnect(this.mount_removed_id);
-        this.connected = false;
-    }
-    destroy() {
-        this.disconnect();
-    }
-}
-
-const Graph = class SystemMonitor_Graph {
-    constructor(extension, width, height) {
-        this.extension = extension;
-        this.menu_item = '';
-        this.actor = new St.DrawingArea({style_class: this.extension._Style.get('sm-chart'), reactive: false});
-        this.width = width;
-        this.height = height;
-        this.gtop = new GTop.glibtop_fsusage();
-        this.colors = ['#888', '#aaa', '#ccc'];
-        for (let color in this.colors) {
-            this.colors[color] = color_from_string(this.colors[color]);
-        }
-
-        let themeContext = St.ThemeContext.get_for_stage(global.stage);
-        themeContext.connect('notify::scale-factor', this.set_scale.bind(this));
-        this.scale_factor = themeContext.scale_factor;
-        let interfaceSettings = new Gio.Settings({
-            schema: 'org.gnome.desktop.interface'
-        });
-        interfaceSettings.connect('changed', this.set_text_scaling.bind(this));
-        this.text_scaling = interfaceSettings.get_double('text-scaling-factor');
-        if (!this.text_scaling) {
-            this.text_scaling = 1;
-        }
-
-        this.actor.set_width(this.width * this.scale_factor * this.text_scaling);
-        this.actor.set_height(this.height * this.scale_factor * this.text_scaling);
-        this.actor.connect('repaint', this._draw.bind(this));
-    }
-    create_menu_item() {
-        this.menu_item = new PopupMenu.PopupBaseMenuItem({reactive: false});
-        this.menu_item.actor.add_child(this.actor);
-        // tray.menu.addMenuItem(this.menu_item);
-    }
-    show(visible) {
-        this.menu_item.actor.visible = visible;
-    }
-    set_scale(themeContext) {
-        this.scale_factor = themeContext.scale_factor;
-        this.actor.set_width(this.width * this.scale_factor * this.text_scaling);
-        this.actor.set_height(this.height * this.scale_factor * this.text_scaling);
-    }
-    set_text_scaling(interfaceSettings, key) {
-        // FIXME: for some reason we only get this signal once, not on later
-        // changes to the setting
-        //log('[System monitor] got text scaling signal');
-        this.text_scaling = interfaceSettings.get_double(key);
-        this.actor.set_width(this.width * this.scale_factor * this.text_scaling);
-        this.actor.set_height(this.height * this.scale_factor * this.text_scaling);
-    }
-}
-
-const Bar = class SystemMonitor_Bar extends Graph {
-    constructor(extension) {
-        // Height doesn't matter, it gets set on every draw.
-        super(extension, extension._Style.bar_width(), 100);
-        this.mounts = extension._MountsMonitor.get_mounts();
-        extension._MountsMonitor.add_listener(this.update_mounts.bind(this));
-    }
-    _draw() {
-        if (!this.actor.visible) {
-            return;
-        }
-        let thickness = this.extension._Style.bar_thickness() * this.scale_factor * this.text_scaling;
-        let fontsize = this.extension._Style.bar_fontsize() * this.scale_factor * this.text_scaling;
-        this.actor.set_height(this.mounts.length * (3 * thickness));
-        let [width, height] = this.actor.get_surface_size();
-        let cr = this.actor.get_context();
-
-        let x0 = width / 8;
-        let y0 = thickness / 2;
-        cr.setLineWidth(thickness);
-        cr.setFontSize(fontsize);
-        for (let mount in this.mounts) {
-            GTop.glibtop_get_fsusage(this.gtop, this.mounts[mount]);
-            const {used, total} = calc_usage(this.gtop);
-            const perc_full = used / total;
-            sm_cairo_set_source_color(cr, this.colors[mount % this.colors.length]);
-
-            let text = this.mounts[mount];
-            if (text.length > 10) {
-                text = text.split('/').pop();
-            }
-            cr.moveTo(0, y0 + thickness / 3);
-            cr.showText(text);
-            cr.moveTo(width - x0, y0 + thickness / 3);
-            cr.showText(Math.round(perc_full * 100).toString() + '%');
-            y0 += (5 * thickness) / 4;
-
-            cr.moveTo(0, y0);
-            cr.relLineTo(perc_full * width, 0);
-            cr.stroke();
-            y0 += (7 * thickness) / 4;
-        }
-        cr.$dispose();
-    }
-    update_mounts(mounts) {
-        this.mounts = mounts;
-        this.actor.queue_repaint();
-    }
-}
-
-const Pie = class SystemMonitor_Pie extends Graph {
-    constructor(extension) {
-        super(extension, extension._Style.pie_size(), extension._Style.pie_size());
-        this.mounts = extension._MountsMonitor.get_mounts();
-        extension._MountsMonitor.add_listener(this.update_mounts.bind(this));
-    }
-
-    _draw() {
-        if (!this.actor.visible) {
-            return;
-        }
-        let [width, height] = this.actor.get_surface_size();
-        let cr = this.actor.get_context();
-        let xc = width / 2;
-        let yc = height / 2;
-        let pi = Math.PI;
-        function arc(r, value, max, angle) {
-            if (max === 0) {
-                return angle;
-            }
-            let new_angle = angle + (value * 2 * pi / max);
-            cr.arc(xc, yc, r, angle, new_angle);
-            return new_angle;
-        }
-
-        // Set the ring thickness so that at least 7 rings can be displayed. If
-        // there are more mounts, make the rings thinner. If the rings are too
-        // thin to have a line height of 1.2 for the labels, shrink the labels.
-        let rings = Math.max(this.mounts.length, 7);
-        let ring_width = width / (2 * rings);
-        let fontsize = this.extension._Style.pie_fontsize() * this.scale_factor * this.text_scaling;
-        if (ring_width < 1.2 * fontsize) {
-            fontsize = ring_width / 1.2;
-        }
-        let thickness = ring_width / 1.5;
-
-        cr.setLineWidth(thickness);
-        cr.setFontSize(fontsize);
-        let r = (height - ring_width) / 2;
-        for (let mount in this.mounts) {
-            GTop.glibtop_get_fsusage(this.gtop, this.mounts[mount]);
-            sm_cairo_set_source_color(cr, this.colors[mount % this.colors.length]);
-            const {used, total} = calc_usage(this.gtop);
-            arc(r, used, total, -pi / 2);
-            cr.stroke();
-            r -= ring_width;
-        }
-        let y = (ring_width + fontsize) / 2;
-        for (let mount in this.mounts) {
-            let text = this.mounts[mount];
-            if (text.length > 10) {
-                text = text.split('/').pop();
-            }
-            cr.moveTo(0, y);
-            cr.showText(text);
-            y += ring_width;
-        }
-        cr.$dispose();
-    }
-
-    update_mounts(mounts) {
-        this.mounts = mounts;
-        this.actor.queue_repaint();
-    }
-}
-
-let TipItem = GObject.registerClass(
-    {
-        GTypeName: 'TipItem'
-    },
-    class SystemMonitor_TipItem extends PopupMenu.PopupBaseMenuItem {
-        _init() {
-            super._init();
-            // PopupMenu.PopupBaseMenuItem.prototype._init.call(this);
-            this.actor.remove_style_class_name('popup-menu-item');
-            this.actor.add_style_class_name('sm-tooltip-item');
-        }
-    }
-);
-const TipMenu = class SystemMonitor_TipMenu extends PopupMenu.PopupMenuBase {
-    constructor(sourceActor) {
-        // PopupMenu.PopupMenuBase.prototype._init.call(this, sourceActor, 'sm-tooltip-box');
-        super(sourceActor, 'sm-tooltip-box');
-        this.actor = new Clutter.Actor();
-        // this.actor.connect('get-preferred-width',
-        //     this._boxGetPreferredWidth).bind(this);
-        // this.actor.connect('get-preferred-height',
-        //     this._boxGetPreferredHeight.bind(this));
-        this.actor.add_child(this.box);
-    }
-    // _boxGetPreferredWidth (actor, forHeight, alloc) {
-    //     // let columnWidths = this.getColumnWidths();
-    //     // this.setColumnWidths(columnWidths);
-    //
-    //     [alloc.min_size, alloc.natural_size] = this.box.get_preferred_width(forHeight);
-    // }
-    // _boxGetPreferredHeight (actor, forWidth, alloc) {
-    //     [alloc.min_size, alloc.natural_size] = this.box.get_preferred_height(forWidth);
-    // }
-    // _boxAllocate (actor, box, flags) {
-    //     this.box.allocate(box, flags);
-    // }
-    _shift() {
-        // Probably old but works
-        let node = this.sourceActor.get_theme_node();
-        let contentbox = node.get_content_box(this.sourceActor.get_allocation_box());
-
-        let sourceTopLeftX = 0;
-        let sourceTopLeftY = 0;
-        let extents = this.sourceActor.get_transformed_extents();
-        let sourceTopLeft = extents.get_top_left();
-        sourceTopLeftY = sourceTopLeft.y;
-        sourceTopLeftX = sourceTopLeft.x;
-        let monitor = Main.layoutManager.findMonitorForActor(this.sourceActor);
-        let [x, y] = [sourceTopLeftX + contentbox.x1,
-            sourceTopLeftY + contentbox.y1];
-        let [cx, cy] = [sourceTopLeftX + (contentbox.x1 + contentbox.x2) / 2,
-            sourceTopLeftY + (contentbox.y1 + contentbox.y2) / 2];
-        let [xm, ym] = [sourceTopLeftX + contentbox.x2,
-            sourceTopLeftY + contentbox.y2];
-        let [width, height] = this.actor.get_size();
-        let tipx = cx - width / 2;
-        tipx = Math.max(tipx, monitor.x);
-        tipx = Math.min(tipx, monitor.x + monitor.width - width);
-        tipx = Math.floor(tipx);
-        let tipy = Math.floor(ym);
-        // Hacky condition to determine if the status bar is at the top or at the bottom of the screen
-        if (sourceTopLeftY / monitor.height > 0.3) {
-            tipy = sourceTopLeftY - height; // If it is at the bottom, place the tooltip above instead of below
-        }
-        this.actor.set_position(tipx, tipy);
-    }
-    open(animate) {
-        if (this.isOpen) {
-            return;
-        }
-
-        this.isOpen = true;
-        this.actor.show();
-        this._shift();
-        this.actor.raise_top();
-        this.emit('open-state-changed', true);
-    }
-    close(animate) {
-        this.isOpen = false;
-        this.actor.hide();
-        this.emit('open-state-changed', false);
-    }
-}
-
-const TipBox = class SystemMonitor_TipBox {
-    constructor(extension) {
-        this.extension = extension;
-        this.actor = new St.BoxLayout({reactive: true});
-        this.actor._delegate = this;
-        this.set_tip(new TipMenu(this.actor));
-        this.in_to = this.out_to = 0;
-        this.actor.connect('enter-event', this.on_enter.bind(this));
-        this.actor.connect('leave-event', this.on_leave.bind(this));
-    }
-    set_tip(tipmenu) {
-        if (this.tipmenu) {
-            this.tipmenu.destroy();
-        }
-        this.tipmenu = tipmenu;
-        if (this.tipmenu) {
-            Main.uiGroup.add_child(this.tipmenu.actor);
-            this.hide_tip();
-        }
-    }
-    show_tip() {
-        if (this.tipmenu)
-            this.tipmenu.open();
-        return GLib.SOURCE_REMOVE;
-    }
-    hide_tip() {
-        if (!this.tipmenu) {
-            return;
-        }
-        this.tipmenu.close();
-        this.stop_out_timer();
-        this.stop_in_timer();
-    }
-    on_enter() {
-        let show_tooltip = this.extension._Schema.get_boolean('show-tooltip');
-
-        if (!show_tooltip) {
-            return;
-        }
-
-        this.stop_out_timer();
-        this.start_in_timer();
-    }
-    on_leave() {
-        this.stop_in_timer();
-        this.start_out_timer();
-    }
-    start_in_timer() {
-        if (!this.in_to) {
-            this.in_to = GLib.timeout_add(
-                GLib.PRIORITY_DEFAULT,
-                this.extension._Schema.get_int('tooltip-delay-ms'),
-                this.show_tip.bind(this),
-            );
-        }
-    }
-    stop_in_timer() {
-        if (this.in_to) {
-            GLib.Source.remove(this.in_to);
-            this.in_to = 0;
-        }
-    }
-    start_out_timer() {
-        if (!this.out_to) {
-            this.out_to = GLib.timeout_add(
-                GLib.PRIORITY_DEFAULT,
-                this.extension._Schema.get_int('tooltip-delay-ms'),
-                this.hide_tip.bind(this),
-            );
-        }
-    }
-    stop_out_timer() {
-        if (this.out_to) {
-            GLib.Source.remove(this.out_to);
-            this.out_to = 0;
-        }
-    }
-    destroy() {
-        this.stop_in_timer();
-        this.stop_out_timer();
-        this.actor.destroy();
-    }
-}
-
-// This class swaps the vertical and horizontal dimensions of a child element,
-// because rotating the child only changes how it is painted, not its geometry.
-// It also moves the bounds on allocation so that the rotated child ends up in
-// the right position.
-const RotateBinLayout = GObject.registerClass(
-    {
-        GTypeName: 'RotateBinLayout'
-    },
-    class SystemMonitor_RotateBinLayout extends Clutter.BinLayout {
-        vfunc_get_preferred_width(container, for_height) {
-            return super.vfunc_get_preferred_height(container, for_height);
-        }
-        vfunc_get_preferred_height(container, for_width) {
-            return super.vfunc_get_preferred_width(container, for_width);
-        }
-        vfunc_allocate(container, box) {
-            const box2 = new Clutter.ActorBox({
-                x1: box.x1,
-                x2: box.x1 + box.y2 - box.y1,
-                y1: box.y2,
-                y2: box.y2 + box.x2 - box.x1,
-            });
-            return super.vfunc_allocate(container, box2);
-        }
-    }
-);
-
-const ElementBase = class SystemMonitor_ElementBase extends TipBox {
-    constructor(extension, properties) {
-        super(extension);
-        this.elt = '';
-        this.elt_short = '';
-        this.item_name = _('');
-        this.color_name = [];
-        this.text_items = [];
-        this.menu_items = [];
-        this.menu_visible = true;
-        this.timeout = null;
-
-        // Maximum value preserved during cooldown period
-        this.graph_scale_max_including_cooldown = 0;
-        this.graph_scale_cooldown_timer_id = null;
-        this.graph_scale_cooldown_delay_minutes = 0;
-
-        Object.assign(this, properties);
-
-        //            TipBox.prototype._init.apply(this, arguments);
-        this.vals = [];
-        this.tip_labels = [];
-        this.tip_vals = [];
-        this.tip_unit_labels = [];
-
-        const Schema = extension._Schema;
-        const Style = extension._Style;
-        const IconSize = extension._IconSize;
-
-        this.colors = [];
-        for (let color in this.color_name) {
-            let name = this.elt + '-' + this.color_name[color] + '-color';
-            let clutterColor = color_from_string(Schema.get_string(name));
-            Schema.connect('changed::' + name, (schema, key) => {
-                this.clutterColor = color_from_string(Schema.get_string(key));
-            });
-            Schema.connect('changed::' + name, () => {
-                this.chart.actor.queue_repaint();
-            });
-            this.colors.push(clutterColor);
-        }
-
-        let element_width = Schema.get_int(this.elt + '-graph-width');
-        if (Style.get('') === '-compact') {
-            element_width = Math.round(element_width / 1.5);
-        }
-        this.chart = new Chart(this.extension, element_width, IconSize, this);
-
-        Schema.connect('changed::background', () => {
-            this.chart.actor.queue_repaint();
-        });
-
-        this.actor.visible = Schema.get_boolean(this.elt + '-display');
-        Schema.connect(
-            'changed::' + this.elt + '-display', (schema, key) => {
-                this.actor.visible = Schema.get_boolean(key);
-            });
-
-        this.restart_update_timer(l_limit(Schema.get_int(this.elt + '-refresh-time')));
-
-        Schema.connect(
-            'changed::' + this.elt + '-refresh-time',
-            (schema, key) => {
-                this.restart_update_timer(l_limit(Schema.get_int(key)));
-            });
-        Schema.connect('changed::' + this.elt + '-graph-width', this.resize.bind(this));
-
-        if (this.elt === 'thermal') {
-            Schema.connect('changed::thermal-threshold',
-                () => {
-                    this.reset_style();
-                    this.restart_update_timer();
-                });
-        }
-
-        this.label = new St.Label({text: _(this.elt_short || this.elt),
-            style_class: Style.get('sm-status-label')});
-        change_text.call(this);
-        Schema.connect('changed::' + this.elt + '-show-text', change_text.bind(this));
-
-        this.menu_visible = Schema.get_boolean(this.elt + '-show-menu');
-        Schema.connect('changed::' + this.elt + '-show-menu', change_menu.bind(this));
-
-        this.label_bin = new St.Bin({child: this.label});
-        const default_layout = this.label_bin.layout_manager;
-        const change_rotate_labels = () => {
-            if (Schema.get_boolean('rotate-labels')) {
-                this.label.set_rotation_angle(Clutter.RotateAxis.Z_AXIS, -90);
-                this.label.add_style_class_name('rotated');
-                this.label_bin.layout_manager = new RotateBinLayout();
-                this.label_bin.y_align = Clutter.ActorAlign.CENTER;
-            } else {
-                this.label.set_rotation_angle(Clutter.RotateAxis.Z_AXIS, 0);
-                this.label.remove_style_class_name('rotated');
-                this.label_bin.layout_manager = default_layout;
-                this.label_bin.y_align = Clutter.ActorAlign.START;
-            }
-        };
-        change_rotate_labels();
-        Schema.connect('changed::rotate-labels', change_rotate_labels);
-
-        this.actor.add_child(this.label_bin);
-        this.text_box = new St.BoxLayout();
-
-        this.actor.add_child(this.text_box);
-        this.text_items = this.create_text_items();
-        for (let item in this.text_items) {
-            this.text_box.add_child(this.text_items[item]);
-        }
-        this.actor.add_child(this.chart.actor);
-        change_style.call(this);
-        Schema.connect('changed::' + this.elt + '-style', change_style.bind(this));
-        this.menu_items = this.create_menu_items();
-
-        this.restart_cooldown_timer();
-        Schema.connect('changed::graph-cooldown-delay-m', () => {
-            this.restart_cooldown_timer();
-        });
-    }
-    /**
-     * Initializes or restarts the graph scale cooldown timer. The graph
-     * scale won't downscale during the cooldown period.
-     *
-     * max - Maximum value to preserve during cooldown
-     */
-    restart_cooldown_timer(max = 0) {
-        if (this.graph_scale_cooldown_timer_id) {
-            GLib.Source.remove(this.graph_scale_cooldown_timer_id);
-        }
-        this.graph_scale_max_including_cooldown = max;
-        this.graph_scale_cooldown_delay_minutes = this.extension._Schema.get_int('graph-cooldown-delay-m');
-        if (this.graph_scale_cooldown_delay_minutes !== 0) {
-            this.graph_scale_cooldown_timer_id = GLib.timeout_add_seconds(
-                GLib.PRIORITY_DEFAULT,
-                this.graph_scale_cooldown_delay_minutes * 60,
-                () => {
-                    this.restart_cooldown_timer();
-                    return GLib.SOURCE_CONTINUE;
-                });
-        }
-    }
-    restart_update_timer(interval = null) {
-        interval = interval || this._lastInterval;
-        if (!interval) {
-            sm_log("Invalid call to restart_update_timer", 'error');
-            return;
-        }
-        if (this.timeout) {
-            GLib.Source.remove(this.timeout);
-        }
-        this.timeout = GLib.timeout_add(
-            GLib.PRIORITY_DEFAULT_IDLE,
-            interval,
-            this.update.bind(this),
-        );
-        this._lastInterval = interval;
-    }
-    tip_format(unit) {
-        if (typeof (unit) === 'undefined') {
-            unit = '%';
-        }
-        if (typeof (unit) === 'string') {
-            let all_unit = unit;
-            unit = [];
-            for (let i = 0; i < this.color_name.length; i++) {
-                unit.push(all_unit);
-            }
-        }
-        for (let i = 0; i < this.color_name.length; i++) {
-            let tipline = new TipItem();
-            this.tipmenu.addMenuItem(tipline);
-            tipline.actor.add_child(new St.Label({text: _(this.color_name[i])}));
-            this.tip_labels[i] = new St.Label({text: ''});
-            tipline.actor.add_child(this.tip_labels[i]);
-
-            this.tip_unit_labels[i] = new St.Label({text: unit[i]});
-            tipline.actor.add_child(this.tip_unit_labels[i]);
-            this.tip_vals[i] = 0;
-        }
-    }
-    //        set_tip_unit: function(unit) {
-    //           for (let i = 0;i < this.tip_unit_labels.length;i++) {
-    //           this.tip_unit_labels[i].text = unit[i];
-    //           }
-    //           }
-    update() {
-        if (!this.menu_visible && !this.actor.visible) {
-            return false;
-        }
-        this.refresh();
-        this._apply();
-        if (this.elt === 'thermal') {
-            this.threshold();
-        }
-        this.chart.update();
-        for (let i = 0; i < this.tip_vals.length; i++) {
-            if (this.tip_labels[i]) {
-                this.tip_labels[i].text = this.tip_vals[i].toString();
-            }
-        }
-        return GLib.SOURCE_CONTINUE;
-    }
-    reset_style() {
-        this.text_items[0].set_style('color: rgba(255, 255, 255, 1)');
-    }
-    threshold() {
-        if (this.extension._Schema.get_int('thermal-threshold')) {
-            if (this.temp_over_threshold) {
-                this.text_items[0].set_style('color: rgba(255, 0, 0, 1)');
-            } else {
-                this.text_items[0].set_style('color: rgba(255, 255, 255, 1)');
-            }
-        }
-    }
-    resize(schema, key) {
-        let width = this.extension._Schema.get_int(key);
-        if (this.extension._Style.get('') === '-compact') {
-            width = Math.round(width / 1.5);
-        }
-        this.chart.resize(width);
-    }
-    destroy() {
-        TipBox.prototype.destroy.call(this);
-        if (this.timeout) {
-            GLib.Source.remove(this.timeout);
-            this.timeout = null;
-        }
-        if (this.graph_scale_cooldown_timer_id) {
-            GLib.Source.remove(this.graph_scale_cooldown_timer_id);
-            this.graph_scale_cooldown_timer_id = null;
-        }
-    }
 }
 
 const Battery = class SystemMonitor_Battery extends ElementBase {
